pyproject.toml
[tool.poetry]
name = "mapistar"
version = "0.4.2"
description = "WebDoctor Api"
authors = ["Jimmy Girardet <ijkl@netc.fr>"]
license = "GPL-3.0+"
readme = "README.rst"
homepage = "https://github.com/jgirardet/mapistar"
repository = "https://github.com/jgirardet/mapistar"
keywords = ['apistar', 'pony', "medical"]


[tool.poetry.dependencies]
python = "^3.6"
pony = "^0.7.3"
psycopg2-binary = "^2.7"
apistar = "^0.5.26"
apistar-jwt = "^0.5.0"
dickens = "^1.0"
pythonthesorimed = "^0.3.0"
simple-settings = "^0.13.0"
apistar-ponyorm = "^0.2.12"
pendulum = "^2.0"
python-dotenv = "^0.8.2"
hug = "^2.4"
marshmallow = "^2.15"
werkzeug = "^0.14.1"


[tool.poetry.dev-dependencies]
# testing
pytest = "^3.6"
pytest-cov = "^2.5"
pytest-pythonpath = "^0.7.2"
pytest-env = "^0.6.2"
pytest-ponyorm = "^0.3.2"
pytest-mock = "^1.10"
mimesis = "^2.0"

# linting
flake8 = "^3.5"
isort = "^4.3"
black = {version = "*", allows-prereleases = true}

# doc
sphinx = "^1.7"
sphinx-rtd-theme = "^0.3.1"
sphinx-autobuild = "^0.7.1"
sphinxcontrib-napoleon = "^0.6.1"
sphinx-autodoc-typehints = "^1.3"
